Moved build_links and req into the view, cut out some redundant exceptions
